requirements.txt
flask
gunicorn
pyhaml
psycopg2
flask-sqlalchemy
flask-migrate
flask-login
authlib
requests
